remove owned references
